refactor(plugins): split validators (runtime) from resolvers (loader)
Item #5a from docs/design/core-boundary-review.md. Within
plugins/skills.py and plugins/agents.py, separate the validator
surface called on the agent-init runtime path from the
manifest-walking resolution code that's only reached from CLI
subcommands and the loader phase of agent boot.

- new agentao/plugins/resolvers/ package: skills.py and agents.py
  hold resolve_plugin_entries / resolve_plugin_agents and all
  private helpers; __init__.py re-exports both and documents the
  split + 5b relocation plan
- agentao/plugins/skills.py and agentao/plugins/agents.py slimmed
  to the single validate_no_external_collisions function each,
  with module docstrings calling out the runtime callers
- delete dead PluginSkillCollisionError (no readers anywhere)
- cli/subcommands.py: 4 imports repointed to plugins.resolvers.*
- 3 test files migrate imports across the new boundary
- runtime callers (skills/manager.py, agents/manager.py) keep
  importing validate_no_external_collisions from the original
  paths — those modules are now validator-only

No back-compat shim left in plugins/skills.py / plugins/agents.py:
resolvers and validators are private package surface (not in
agentao.plugins.__all__) and only first-party callers were touched.

Tests: 2549 passed, 2 skipped. No regressions.
