+++ tests/test_plugin_skills.py
@@ -13,10 +13,8 @@ from agentao.plugins.models import (
     PluginManifest,
     PluginSkillEntry,
 )
-from agentao.plugins.skills import (
-    resolve_plugin_entries,
-    validate_no_external_collisions,
-)
+from agentao.plugins.resolvers.skills import resolve_plugin_entries
+from agentao.plugins.skills import validate_no_external_collisions
 
 
 # ======================================================================
